test scripts.js operation
